change packages, add usecase
