+++ app/src/main/AndroidManifest.xml
@@ -11,8 +11,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.HeadwayTestApp"
-        android:name = ".di.HeadwayTestApp" >
-        <activity android:name=".MainActivity"
+        android:name = "com.example.headwaytestapp.utils.di.HeadwayTestApp" >
+        <activity android:name=".presentation.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
