+++ app/src/main/java/com/example/headwaytestapp/data/dao/RepositoryDAO.kt
@@ -1,4 +1,4 @@
-package com.example.headwaytestapp.dao
+package com.example.headwaytestapp.data.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
