+++ app/src/main/java/com/example/headwaytestapp/presentation/authorization_view/UiStateManager.kt
@@ -1,4 +1,4 @@
-package com.example.headwaytestapp.authorization_view
+package com.example.headwaytestapp.presentation.authorization_view
 
 import com.google.firebase.auth.OAuthProvider
 
@@ -6,6 +6,6 @@ sealed class UiStateManager {
     object InProgress : UiStateManager()
     object SuccessLogIn : UiStateManager()
 
-    data class ErrorLogIn(val message: String?) : UiStateManager()
+    data class ErrorLogIn(val message: String) : UiStateManager()
     data class OpenLogIn(val provider: OAuthProvider.Builder) : UiStateManager()
 }
