+++ app/src/main/res/layout/repository_item.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="repositoryItem"
-            type="com.example.headwaytestapp.dao.Repository" />
+            type="com.example.headwaytestapp.data.dao.Repository" />
 
         <import type="android.view.View" />
     </data>
